gnu aanpassingen

+++ 3. Matlab/Tests/2. Groupsize - perc Reconstruction/TestSetupLinprog.m
@@ -13,7 +13,7 @@ result(pos_idx) = 1;
 % m = measurement size
 % p = P(A_i,j == 1) = de kans dat een individu in een test zit
 p = 1/k;
-A = logical(sprand(m,n,p));
+A = double(logical(sprand(m,n,p)));
 
 % b = undersampled measurement, b = A v result
 b = double(logical(A*result));
